fix fredfof and vcraft and craft2fil - possibly need to do more fixing as things not quite right

+++ python/fredfof.py
@@ -66,7 +66,7 @@ def load(fin, values):
         else: # ignore it
             pass
 
-    return np.array(d)
+    return np.array(d, dtype=np.float32)
 
 def fof_file(fin, values):
     d = load(fin, values)
